-fixed bug in departments where the schedules file was raising false positive exception -added extra chek while deleting lab for making sure all schedules associated with that lab are deleted -added input sanitizing for converting lab name to upper case -other small fixes

+++ departments/DepartmentActions.php
@@ -167,12 +167,20 @@ try
 
             if($deletepath){
 
+                $count = count(array_diff(glob($deletepath.'/*'), array('.', '..','.schedules')));
+                if($count != 0){
+                    throw new Exception("There is atleast one LAB assigned to these department, please delete the LABs or assign them to other department");
+                }
+
+                //delete schedules file assuming its empty, i know its not safe
+                unlink($deletepath."/.schedules");
                 $status = rmdir($deletepath);
+
                 if($status){
                     $jTableResult['Result'] = "OK";
                 }else{
                     $jTableResult['Result'] = "ERROR";
-                    $jTableResult['Message'] = "There is atleast one LAB assigned to these department, please delete the LABs or assign them to other department";
+                    $jTableResult['Message'] = "Error occurred while deleting department, directory might not be empty";
                 }
             }else{
                 $jTableResult['Result'] = "ERROR";
